refactor(spi): point graph, retrieval, and ingestion code at com.embabel.dice.spi
Update this branch's new code (graph projectors, graph/discovery query types,
the ingestion artifact model, and their tests) to import the policy SPIs
(TrustScorer, AuthorityResolver/AuthorityTier and friends, ConflictType) from
the com.embabel.dice.spi package they now live in.
