+++ dice/src/main/kotlin/com/embabel/dice/projection/graph/LlmGraphProjector.kt
@@ -19,10 +19,10 @@ import com.embabel.agent.api.common.Ai
 import com.embabel.agent.core.AllowedRelationship
 import com.embabel.agent.core.DataDictionary
 import com.embabel.common.ai.model.LlmOptions
-import com.embabel.dice.common.AuthorityResolver
+import com.embabel.dice.spi.AuthorityResolver
 import com.embabel.dice.common.Relation
 import com.embabel.dice.common.Relations
-import com.embabel.dice.common.StructuralAuthorityResolver
+import com.embabel.dice.spi.StructuralAuthorityResolver
 import com.embabel.dice.proposition.MentionRole
 import com.embabel.dice.proposition.ProjectionFailed
 import com.embabel.dice.proposition.ProjectionResult
