+++ dice/src/test/kotlin/com/embabel/dice/projection/graph/EdgeAuthorityProjectionTest.kt
@@ -20,8 +20,8 @@ import com.embabel.agent.core.DataDictionary
 import com.embabel.agent.rag.model.NamedEntityData
 import com.embabel.agent.rag.service.NamedEntityDataRepository
 import com.embabel.agent.rag.service.RelationshipData
-import com.embabel.dice.common.AuthorityTier
-import com.embabel.dice.common.FixedAuthorityResolver
+import com.embabel.dice.spi.AuthorityTier
+import com.embabel.dice.spi.FixedAuthorityResolver
 import com.embabel.dice.common.Relations
 import com.embabel.dice.proposition.EntityMention
 import com.embabel.dice.proposition.MentionRole
